fix: remove old glofas station endpoints
