+++ services/API-service/src/app.module.ts
@@ -8,7 +8,6 @@ import { CountryModule } from './api/country/country.module';
 import { DisasterModule } from './api/disaster/disaster.module';
 import { EapActionsModule } from './api/eap-actions/eap-actions.module';
 import { EventModule } from './api/event/event.module';
-import { GlofasStationModule } from './api/glofas-station/glofas-station.module';
 import { LinesDataModule } from './api/lines-data/lines-data.module';
 import { MetadataModule } from './api/metadata/metadata.module';
 import { NotificationModule } from './api/notification/notification.module';
@@ -34,7 +33,6 @@ import { TypeOrmModule } from './typeorm.module';
     EventModule,
     MetadataModule,
     AdminAreaModule,
-    GlofasStationModule,
     AdminAreaDynamicDataModule,
     DisasterModule,
     AdminAreaDataModule,
